// frontend/src/app/core/api/model/accelerated-promotion-request.ts
/**
 * Training school system Backend API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface AcceleratedPromotionRequest { 
    name: string;
    code: string;
    startDate: string;
    endDate: string;
    fee: number;
    trainingId: number;
    numberOfHours: number;
    level?: string;
}